pyproject.toml
[project]
name = "delto-tecnica-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Cristian Rodríguez", email = "cristian2x1002@gmail.com" }]
dependencies = [
    "aiohttp>=3.11.18",
    "aiosqlite>=0.21.0",
    "pydantic>=2.11.4",
    "pyowm>=3.3.0",
    "python-telegram-bot>=22.0",
]

[project.scripts]
dev = "main:dev"

[tool.uv]
package = true

[tool.ruff]
line-length = 80

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
